+++ tests/models/test_rl/integration/test_policy_models.py
@@ -25,7 +25,7 @@ class TestPolicyModels(TestCase):
             gpus=0,
             max_steps=100,
             max_epochs=100,  # Set this as the same as max steps to ensure that it doesn't stop early
-            val_check_interval=1000,  # This just needs 'some' value, does not effect training right now
+            val_check_interval=1,  # This just needs 'some' value, does not effect training right now
             fast_dev_run=True
         )