+++ tests/models/test_rl/integration/test_value_models.py
@@ -32,7 +32,7 @@ class TestValueModels(TestCase):
             gpus=self.hparams.gpus,
             max_steps=100,
             max_epochs=100,  # Set this as the same as max steps to ensure that it doesn't stop early
-            val_check_interval=1000,  # This just needs 'some' value, does not effect training right now
+            val_check_interval=1,  # This just needs 'some' value, does not effect training right now
             fast_dev_run=True
         )